moved scenario-cli to meta crate
