+++ framework/meta/Cargo.toml
@@ -26,6 +26,7 @@ template-test-released = []
 
 [dependencies]
 clap = { version = "4.4.7", features = ["derive"] }
+tokio = { version = "1.24", features = ["full"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 toml = { version = "0.8.6", features = ["preserve_order"] }
@@ -39,13 +40,20 @@ copy_dir =  "0.1.2"
 pathdiff = "0.2.1" 
 common-path = "1.0.0" 
 
+[dependencies.multiversx-sc-meta-lib]
+version = "=0.50.3"
+path = "../meta-lib"
+
 [dependencies.multiversx-sc]
 version = "=0.50.3"
 path = "../base"
 features = ["alloc", "num-bigint"]
 
-
-[dependencies.multiversx-sc-meta-lib]
+[dependencies.multiversx-sc-scenario]
 version = "=0.50.3"
-path = "../meta-lib"
+path = "../scenario"
+
+[dependencies.multiversx-sdk]
+version = "=0.4.1"
+path = "../../sdk/core"
