+++ framework/meta/src/cmd.rs
@@ -3,6 +3,7 @@ pub mod info;
 pub mod install;
 pub mod local_deps;
 pub mod print_util;
+pub mod retrieve_address;
 pub mod scen_test_gen;
 pub mod template;
 pub mod test;
