+++ framework/scenario/Cargo.toml
@@ -33,10 +33,6 @@ clap = { version = "4.4.7", features = ["derive"] }
 tokio = { version = "1.24", features = ["full"] }
 unwrap-infallible = "0.1.5"
 
-[[bin]]
-name = "sc-scenario"
-path = "src/main.rs"
-
 [features]
 run-go-tests = []
