--- framework/scenario/src/main.rs
@@ -1,4 +1,0 @@
-#[tokio::main]
-async fn main() {
-    multiversx_sc_scenario::standalone::cli_main().await;
-}
