+++ framework/scenario/src/standalone/mod.rs
@@ -1,5 +1,3 @@
-mod account_tool;
-mod scenario_cli;
+pub mod account_tool;
 
 pub use account_tool::retrieve_account_as_scenario_set_state;
-pub use scenario_cli::cli_main;
